refactor: extract ProjectSerializer from RfSimulatorApp
