+++ app/include/app.h
@@ -26,6 +26,7 @@
 #include "pfb_channelizer_engine.h"
 #include "pfb_channelizer_widget.h"
 #include "pfb_view_manager.h"
+#include "project_serializer.h"
 #include "session_state.h"
 #include "signal_generator_engine.h"
 #include "signal_generator_widget.h"
@@ -121,6 +122,9 @@ class RfSimulatorApp {
     // Declared after m_components so the manager (and its widget references to
     // engines) is destroyed before the engines themselves.
     PFBViewManager m_pfb_views;
+    // Owns .rfsim save/load/new; declared after m_graph_widget and m_pfb_views
+    // so it is destroyed before them (it holds references to both).
+    std::unique_ptr<ProjectSerializer> m_serializer;
     int m_next_component_id = 100;
     PendingAction m_pending_action = PendingAction::None;
     bool m_show_unsaved_dialog = false;
